Fixed alltoallv bug. The changes were taken from branch kpath-unrolled-alltoallv-debug
The bug was that there were some ranks that sends nothing to some other ranks. In this case the find_nbr logic was not setting any value in scount and rcount arrays, so displacements got screwed up.

To overcome this I thought I could send zero count with alltoallv, but testing it showed the behavior is undefined, so instead now sets as send 1 dummy data item. The good thing is we never have to set or read this dummy value because offsets are correctly setup so vertices will read whatever they are supposed to send or recv